pyproject.toml
[project]
name = "src"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "matplotlib>=3.10.1",
    "pandas>=2.2.3",
    "pyodbc>=5.2.0",
    "python-dotenv>=1.0.1",
    "seaborn>=0.13.2",
    "streamlit>=1.43.2",
]
